pyproject.toml
[project]
name = "ai-guide"
version = "1.0.0"
description = "AI Guide - Umfassender Leitfaden zur optimalen Prompt-Formulierung für KI-Modelle"
authors = [
    {name = "A. Koc", email = "ak@ak-pro.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup",
]

dependencies = [
    "Flask==3.0.0",
    "Werkzeug==3.0.1",
    "Jinja2==3.1.2",
    "MarkupSafe==2.1.3",
    "itsdangerous==2.1.2",
    "click==8.1.7",
    "blinker==1.6.3",
    "markdown==3.5.1",
    "Pygments==2.17.2",
    "PyJWT==2.8.0",
    "gunicorn==21.2.0",
    "flask-cors==4.0.0",
    "flask-limiter==3.5.0",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-flask==1.3.0",
    "black==23.11.0",
    "flake8==6.1.0",
    "bandit==1.7.5",
    "sentry-sdk[flask]==1.38.0",
    "cryptography>=42.0.0",
]

[project.urls]
Homepage = "https://ai-guide.ak-pro.com"
Repository = "https://github.com/ak-hosting/ai-guide"
Documentation = "https://ai-guide.ak-pro.com"
Issues = "https://github.com/ak-hosting/ai-guide/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 